Move useWantedTransaction to hooks
