+++ src/app/hooks/useWantedTransaction.ts
@@ -1,0 +1,33 @@
+import { RuntimeTransaction, RuntimeTransactionList } from '../../oasis-nexus/api'
+
+type TransactionSelectionResult = {
+  wantedTransaction?: RuntimeTransaction
+  warningMultipleTransactionsSameHash?: boolean
+}
+
+/**
+ * Find the wanted transaction, in case there are more.
+ *
+ * Normally we want the successful one. If there is none, then the latest.
+ */
+export function useWantedTransaction(
+  transactionsList: RuntimeTransactionList | undefined,
+): TransactionSelectionResult {
+  const transactions = transactionsList?.transactions ?? []
+
+  if (!transactions.length) {
+    // Loading or error
+    return {}
+  } else if (transactions.length === 1) {
+    return {
+      wantedTransaction: transactions[0],
+    }
+  } else {
+    const successfulOne = transactions.find(transaction => transaction.success)
+    const latestOne = transactions.sort((a, b) => b.round - a.round)[0]
+    return {
+      warningMultipleTransactionsSameHash: true,
+      wantedTransaction: successfulOne ?? latestOne,
+    }
+  }
+}
